chore: move benchmark script and convert it to use the actual modulerunner
